backend/requirements.txt
Django==4.2.4
psycopg2-binary==2.9.7
Pillow==10.0.0
flake8==6.1.0
djangorestframework==3.14.0
PyYAML==6.0.1
requests==2.31.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
django-shortuuidfield==0.1.3
django-rest-passwordreset==1.3.0
python-dotenv==1.0.0
drf-spectacular==0.27.0
coverage==7.3.3
black==23.11.0
